Refactor Dependabot and Renovate packages to changelog/sources
